CipherSharp.Ciphers.Tests/Substitution/MultiBeaufortTests.cs
using CipherSharp.Ciphers.Substitution;
using System;
using Xunit;

namespace CipherSharp.Tests.Ciphers.Substitution
{
    public class MultiBeaufortTests
    {
        [Fact]
        public void Encode_BasicParameters_ReturnsCipherText()
        {
            // Arrange
            string text = "helloworld";
            string[] keys = new string[2] { "test", "key" };
            MultiBeaufort multiBeaufort = new(text, keys);
            // Act
            var result = multiBeaufort.Encode();

            // Assert
            Assert.Equal("YERCZQGIWI", result);
        }

        [Fact]
        public void Decode_BasicParameters_ReturnsPlainText()
        {
            // Arrange
            string text = "YERCZQGIWI";
            string[] keys = new string[2] { "test", "key" };
            MultiBeaufort multiBeaufort = new(text, keys);

            // Act
            var result = multiBeaufort.Decode();

            // Assert
            Assert.Equal("HELLOWORLD", result);
        }

        [Fact]
        public void NewInstance_NullMessage_ThrowsArgumentException()
        {
            // Arrange
            string message = null;
            string[] keys = { "test" };
            // Act

            // Assert
            Assert.Throws<ArgumentException>(() => new MultiBeaufort(message, keys));
        }

        [Fact]
        public void NewInstance_NullKeys_ThrowsArgumentNullException()
        {
            // Arrange
            string message = "test";
            string[] keys = null;
            // Act

            // Assert
            Assert.Throws<ArgumentNullException>(() => new MultiBeaufort(message, keys));
        }
    }
}
